ci: scylla tests happy path
